requirements.txt
streamlit
PyPDF2
numpy
faiss-cpu
sentence-transformers
transformers
torch
